Fix various bugs found via nestest, implement undocumented opcodes.
This commit fixes a few bugs found via nestest, following along at
http://www.qmtpro.com/~nes/misc/nestest.log and using
https://wiki.nesdev.com/w/index.php/CPU_unofficial_opcodes as reference.
Currently, the emulator is not in the proper startup state (we jump to 0xC000
at start, in accordance with nestest startup instructions). Not all undocumented
opcodes have been implemented yet, so the emulator currently panics on the first
undocumented one it finds.