Add further PPU rendering logic, and clean up misc. bugs.
This commit improves the quality of PPU rendering logic, (mostly)
properly rendering backgrounds, and rendering sprites as a monocolored
box. Many bugs still remain and this code is primarily untested, in its
current state. The primary bug to be fixed immedietly is that it appears
the PPU attribute table values are being incorrectly read, causing boxes
of incorrect colors being rendered on the background.