PPU bugfixes and misc. cpu instruction timing fixes.
This commit includes a large number of fixes to the PPU to make it more
functional. This commit also includes a few miscellaneous fixes to CPU
instruction timing. Fine_x implementation is a bit hacked together right
now, so all background tiles will snap to the nearest 8 pixels. Not
ideal, but it makes super mario playable. This commit contains a great
deal of logging at the moment. Not ideal, but I can't be bothered to
remove it.